refactor(datagrid-web): split module
